Add yearly and weekly archive generation and layout
Add preprocessor Rules that will generate the archives.

Add a special layout for archives and render correct layout for
archives.

Add functions that generate the archives. This code is copied and
adapted from the CCCS [0] website. The original author is Skyr [1] and
their code is released under the CC BY-SA 3.0 Licence [2].

[0]:
https://github.com/cccs/cccs-website/blob/master/lib/yearlyarchive.rb

[1]:
https://github.com/Skyr

[2]:
http://creativecommons.org/licenses/by-sa/3.0/