Create links to previous and next items
In a new preprocessor function add links to previous and next items to
articles, yearly archives and weekly archives.

NOTICE: This generation is very slow. When some new features are tested it could be better to comment it temporary.

With this I can remove previous (and duplicate) creation code out of the
archives files.

I also move the rendering to the default layout, thus adding consistency
to those links. In the `head` of every page I add meta links to the next
and the previous items.