Generate DirectoryIndex from Nanoc configuration
Nanoc uses the configuration ‘index_filenames’ in nanoc.yaml to specify
which files are – well – index files. That affects the method path() on
Items. Normaly this is just ‘index.html’ but I added ‘index.xml’ for
pretty feed URLs.

Problem: Specifying a DirectoryIndex in the htaccess and adding filenames to the configuration requires duplicate effort.

Solution: Generate DirectoryIndex from ‘index_filenames’ in the configuration.