Improve generated URLs for tag pages
Ruby doesn’t change the case for uppercase umlauts. This results in bad
URLs. As an example:

    irb > "Überwachung".downcase
    => "Überwachung"

So I had to add a few more chars to the regular expression, so that no
quotation marks and umlauts will end in the tag pages. Simple ignoring
the case in the regular expression results in “ss” also being removed by
gsub, which is quite annoying.