Make sure Snippets can be used in the layout
A few texts in the layout are used as link titles. They shouldn’t
contain linebreaks, HTML code and whitespace at the end or the
beginning. Thus I add the new helper function “snippet_for_title”, that
loads a snippet and prepares its content to be used in a links title.
This also requires the inclusion of Nanoc::Helpers::Text.

Snippets containing ERB and being used in the layout, have to be
compiled before the all other items. To take care of that, I add a rule
that compiles snippets before anything else.

* Add function to use snippets as title
* Compile Snippets first