SASS: Load color variables from a single file
Define all used color variables in the new SASS partial “colors”. That
partial is loaded after the CSS reset but before anything else. The
other partials (“typography” and “menu”) are modified to use the
variables.

This introduces a new design. A dark background with white (well, grey)
text and green links. I like it and I think it’s more pleasant to look
at in the night.

* Use variables for colors.
* Implement a dark colorscheme.