SCSS: Fix multiple spacing issues
Remove some TODO comments in headlines. I don’t know how to resolve that
and I don’t want to invest time in that.

Increase top and bottom margin of <aside> to match the bottom margin of
<figcaption> (added in 769635d799cb). Also increase line-height. That
makes it easier to read the white font on black background.

Add some spacing to iframes (embedded videos on YouTube) and try to make
them a little bit more responsive. This doesn’t work perfect, but is
okay with much effort.