Improve colors for input text fields
The placeholder has been nearly invisible. After changing the inputs
text color to the normal text color, it’s more readable. Adding a
background color makes clear, how big the field is.

When the field is focused, the borders colors are changed. That way, a
user has a better feedback, when she clicks an input field.

* Change border, background and text colors on input fields.