Use git data as a fallback (e.g. date of commit)
Some data is saved in the yaml part of the content files and in the git
repository. That is redundant and leads to imprecise data (in the yaml
part and thus in the published data).

Now, the publishing date of an article is defined as the first commit to
branch “master”. That way, it’s possible to just don’t declare the
“created_at” attribute of a post. However, if the attribute is declared
in the content file, it won’t be overwritten.

Some things are required, that all helpers still work properly.

0. We don’t use functions, but attributes. Most helpers use the data
   from the yaml part (e.g. “item[:created_at]”) and require the
   presence of those fields.

1. The data from git is loaded as a preproccessor rule. This ensures,
   that the data is available before the helpers are loaded.

Thanks to the authors of the awesome git library for ruby.

https://github.com/schacon/ruby-git