Use git data to get an author fallback
When there isn’t an author associated to the page, the author of the
first commit to the master branch is choosen as a fallback. The used
field is `:author_name` as this is used by Nanocs atom feed generation.

Also display the author on the homepage with the date, so this change
makes any sense.