Remove Markdown flavour for headlines
Headlines with monospaced font and preceeding #s, is a funny idea, but I
guess it doesn’t look very well. That is why I removed them.

Now the headlines use the default font (which should be Ubuntu) and only
have the different sizes and margins above and below as form of
contrast. It seems, that the page looks a little bit cleaner and calmer
now. I like it.