Add linked date information in the article partial
Keeping the date information in the footer of a single page, would be
redundant. Also displaying the date as paragraph below the headline,
looks cleaner than the lonely sentence in the footer. That’s why I
removed the date information there.

Without the date information in the footer, only blog articles will have
dates attached. That’s another reasonable advantage.