Generate categories from tags
The function `all_items_with_category` now uses the current item (@item)
as the category. That way, only categories can fetch the items in that
category. This is very restrictive.

`all_items_with_category` returns an empty array, when the items kind is
not “category” or when the category doesn’t have any tags.

Every article in `sorted_articles` will be categorized, as long as it
has some tags. No article should be twice in the same category. An
article will be added if one of it’s tags matches one of the categories
tags.

Github-Issue: https://github.com/rosetree/plasisent.org/issues/10