Remove the font selections
The fonts I specified in CSS haven’t been a choice; they were the
absence of a choice.

I like the Ubuntu typeface. But that was the only reason I used it. That
is not very good. I should try to think about which font I should use in
my blog and *why*. As long as I don’t do that, I will be punished with
default fonts.