Add unobstrusive style to <footer>
The footer element contains metadata, that’s not so imortant and thus
shouldn’t pull focus from the main content. That’s why I set opacity to
0.6 (luminosity is 8.13 according to WCAG 2).

An hover, the footer element should receive more focus, so the opacity
is set to 0.9. With side padding of 3rem and a border on the side, the
footer looks more compact.