Don’t use <article> in permalink pages
In permalink pages, there’s only a single article. There’s no need to
specify this as an <article>. However on the homepage are more than one
articles, so using <article> scoping makes sense.

This improves the document outline. The articles headline of a permalink
page, is a first level headline. On the homepage, the firstlevel
headline is the sites name, and the headlines of all articles are
second-level headlines beneath the sites-name.

Homepage:
* Plasisent
  * Article 0
  * Article 1
  * Article 2

Permalink:
* Plasisent
* Article

This can be checked using the HTML 5 Validator [0].

[0]: https://validator.w3.org/nu/