Add generated categories to their articles
Every article, that is collected in a category, receives an information
about that category. This NanocItem has to be changed. NanocItems can
only be changed in the preprocessor.

This adds a new parameter to all_items_with_category, that contains an
item. The item is supposed to be of kind "category".

* Show all categories of an article in the default layout.