Create atom feed for categories
The new function `create_category_feed` will be called in the
preprocessing function for categories. For each category it adds a new
Nanoc::Item to the @items; the feed for this category.

The path of the feeds is: “/feed/category-path/”. That way, every feed
on Plasisent is accessible in plasisent.org/feed/

Those feeds are generated with the `atom_feed` function from Nanoc.

See also:
* http://nanoc.ws/docs/api/Nanoc/Helpers/Blogging.html#atom_feed-
instance_method