Add a tl;dr note to articles
That note could be used as a small introduction to the article, that
summarizes my main issue.

With this commit I add an <aside> element containing that note with a
link to tweet it with the URL of the current article. I also use that
note to add a meta description to the html head.

Technically I use the the Capturing Helper from Nanoc.

see also: https://nanoc.ws/docs/api/Nanoc/Helpers/Capturing.html