[Fix] Add fallback date on non published posts
Because I used git data, there was no need to add created_at data at the
top of new articles. But on drafts, that had no commit yet, that
generated an error because, some used features (like the atom_feed) need
a generation date.

Now I add an fallback date (which is set to today), when there is no
creation date set and there is no commit on this file. Thus I am able to
write drafts, that don’t have a created_at attribute.