Update wording on the homepage
The title is more descriptive. It’s my blog, not a blog about a specific
topic. As there is a small navigation on the top of every page, I think
we don’t need the <nav> here anymore. Also the homepage <nav> links to
pages, that are not well maintained.