Show all kind of posts on the homepage
My <aside> notes for featured blogs and articles, should only be shown
after articles and not after other posts.

The count of articles reflects the count of articles for now. That means
if I write a hundred notes between two articles, all of them will be
listed. That needs surveillance.