Optimize <footer> style
Opacity should not change on hover, because this is annoying on touch
screens. The <footer> is different and we show this with the large
margin-top.

Use the border style, we already know from article.note. Add a padding
top of 1.5rem, a size we use widely on the side. Thus this commit makes
<footer> more consistent and includes it more into the blog.