Remove categories and tags from the output and build
Additionally to removing the categories and tags from the compilation
rules, this commt adds a note in the error 404 page, so users get an
info. Also all occurences in the layout and in some texts are removed.

This feature is disabled, because it doesn’t seem to be used very often
(aside from bots of course) and it keeps me from writing. Actually I
spend this evening creating this commit and *not* writing a new post.