Add ExpiresByType in .htaccess to image files
Images with an URL won’t change in a way, that makes a new fetch
necessary. To make sure, a browser will know that it can savely cache an
image file for a longer period, the server sends an Expires information
in the response headers.

Expires directives shall only be executed, if mod_expires is enabled. Do
this with administrative rights (probably as root) with the following
commands

    $ a2enmod expires
    $ service apache2 restart

see more
* https://httpd.apache.org/docs/2.2/mod/mod_expires.html
* https://css-tricks.com/snippets/htaccess/set-expires/