Mount all static items into “/” instead of “/static/”
Binary items are skipped from compilation, as they have been before. But
we have to adapt compiling and routing for php files (→ the only php
file available).

This doesn’t change routing in the output. That’s because all items with
a “/static/” prefixed identifier have been routed to “/” before.