Add critical CSS to the default layout
These few lines of CSS are used as inline styles in the <head> of the
document. Using inline styles, makes sure a user gets a very simple
styling without a seperate HTTP request (for the stylesheet).

Ordering is important in this case. The critical CSS should be placed
*before* the stylesheets, so their settings can be overwritten.

This is inspired by an npm.js package criticalcss [0], but manually
without adding another dependency. Mat “Wilto” Marquis recommended it in
his talk about “Smaller, Faster Websites” [1].

[0]: https://www.npmjs.com/package/criticalcss
[1]: https://bocoup.com/weblog/smaller-faster-websites