Add RewriteRule for links to /yyyy/mm/dd/slug/
If the requested file does not exist, this rewrite rules will just drop
the month and day, to match my permalink structure.

Plasisent needs this, because some of my articles link to some other
articles with the old permalink scheme. This is pretty bad and should be
changed. By using this rewrite rule, I can also handle external
requests.