Override the default selection colors
Use the link color as background and the normal background as
foreground. Looks good to me.

The Mozilla Developer Network (MDN) states, that this might not work for
everybody, because it’s no longer in the CSS3 spec. I don’t really care.
If a browser doesn’t support this small feature the user won’t miss
anything, the site still works. But it’s an eye candy for everyone else.

⋄ see also:
· https://css-tricks.com/almanac/selectors/s/selection/
· https://developer.mozilla.org/en-US/docs/Web/CSS/::selection