Add CSS rules for the <mark> element
Use the text selection colors (this shouldn’t be a huge problem, because
<mark> should be uses sparingly in my opinion). Except for blockquotes,
use secondary text color as the background color there.

I believe the <mark> element is useful for authors, so they can
highlight what is important for them. As an example a writer is
able to refer to a specifc part of a quotation.

⋄ see also:
· https://www.w3.org/wiki/HTML/Elements/mark
· https://www.w3.org/TR/html51/semantics.html#the-mark-element