Move contentinfo <footer> into <main>
Put the content where it belongs to → Information about the current page
(shortlink, link to github, comment to email, changes) belongs to the
main content of the page. And as everything is a link, we can use a
navigation list (without specific ordering).

Also use a better title then before.