Add site wide <footer> at the bottom of every page
This is the place where a short description about the blog, a site wide
<nav> and “follow me” links are appropriate: after the main content. Now
the page reflects semantically what’s promised in the <title>.

Every page has a title like this:
<title>Headline of Page Content · Plasisent</title>

It’s also what I normally do. First I read the content of the page that
was linked somewhere. After I finished I decide whether or not it’s
worthwhile to learn more about the blog or to follow the author.

⋄ read more
· https://vasilis.nl/nerd/where-should-the-logo-be/
· https://vasilis.nl/nerd/where-should-the-navigation-be/