Enhance <nav> style with multiple lines
The list elements need some inline space (line-height) but not any
margins. Using display inline makes sure that long items can be splitted
between words and don’t need to be on the same line.

We also add some margin above a navigation, so that they can be placed
as an example below a normal paragraph.