Move everything from static/ to content/
This follows an advide from the extended upgrade guide to Nanoc 4. Just
to move all these files is a lot easier, compared to following the each
of the steps in this guide.

⋄ http://nanoc.ws/doc/nanoc-4-upgrade-guide/#extended-upgrade-guide

“A final improvement would be to move the contents of the static/
directory into content/.”

I used “meld” to check whether the output is still working as I expect
and as far as I could see, it does. So this seems to be the easiest way
to handle my static content dir.

⋄ http://meldmerge.org/